Move `urlWithQueryString` into `http`
